Added functionality to check the 7 days forecast

+++ weatherApp/views.py
@@ -1,3 +1,5 @@
+import datetime
+
 import requests
 from django.shortcuts import render, redirect
 
@@ -8,7 +10,6 @@ from .forms import CityForm
 
 def weather_view(request):
     url = 'https://api.openweathermap.org/data/2.5/weather?q={}&units={}&appid=110a6ca59b063163742ebf7ff21d3e04'
-    city = 'Shanghai'
 
     err_msg = ''
     user_message = ''
@@ -49,9 +50,11 @@ def weather_view(request):
 
         city_weather = {
             'city': city.name,
-            'temperature': r['main']['temp'],
+            'temperature': round(r['main']['temp']),
             'description': r['weather'][0]['description'],
             'icon': r['weather'][0]['icon'],
+            'lon': r['coord']['lon'],
+            'lat': r['coord']['lat'],
         }
         weather_data.append(city_weather)
 
@@ -61,5 +64,27 @@ def weather_view(request):
 
 
 def delete_city(request, city_name):
-    City.objects.get(name=city_name).delete()
-    return redirect('home')
+    if request.method == 'POST':
+        City.objects.get(name=city_name).delete()
+        return redirect('home')
+
+    return render(request, 'weatherApp/weather_confirm_delete.html', {'city_name': city_name})
+
+
+def seven_days_weather(request, lon, lat, city_name):
+    url = 'https://api.openweathermap.org/data/2.5/onecall?lon={}&lat={' \
+          '}&units=metric&appid=110a6ca59b063163742ebf7ff21d3e04&exclude=minutely,hourly '
+
+    r = requests.get(url.format(lon, lat)).json()
+
+    daily_info = []
+    for day in r['daily']:
+        daily_info.append({'day_of_week': datetime.datetime.fromtimestamp(day['dt']).strftime('%A'),
+                           'date': datetime.datetime.fromtimestamp(day['dt']).date(),
+                           'temp_day': round(day['temp']['day']),
+                           'temp_night': round(day['temp']['night']),
+                           'weather': day['weather'][0]['main'],
+                           'description': day['weather'][0]['description'],
+                           'icon': day['weather'][0]['icon']})
+
+    return render(request, 'weatherApp/seven_days_weather.html', {'daily_info': daily_info, 'city_name': city_name})
